chainparams: reverse checkpoint hash bytes

+++ BRChainParams.h
@@ -61,18 +61,18 @@ static const char *BRTestNetDNSSeeds[] = {
 // blockchain checkpoints - these are also used as starting points for partial chain downloads, so they must be at
 // difficulty transition boundaries in order to verify the block difficulty at the immediately following transition
 static const BRCheckPoint BRMainNetCheckpoints[] = {
-    {     0, "aec9203efaf5b35bdfc136477468830da578a1f5b21f4c59cf898dbb3fb17c33", 1512973198, 0x1e10024c },
-    {     1, "8852898d57539ff7e065b16149d2097a37b41617b963d6002a7a00ecd1eeadc8", 1514518477, 0x1e10024c },
-    {  2608, "22722429354af4f4ef8b9de55f0825611aa348d98fcb05d71e3c3adfaab4b94c", 1514547220, 0x1e10024c },
-    {  5216, "9726ea3e333ca64e6cffa5fa9fe0e27338634890b36ad09b75c9c04eee59d423", 1514586552, 0x1e040093 },
-    { 10432, "8c53c852c42c4ae1fad07c12cf8d259baf3ca873d24cac84cdf679d92dc6dcb6", 1514770971, 0x1d100240 },
-    { 20864, "12e76ce3468f0c2fa8d53458ac69967faf88cc1e1d43641cf3f2afb55d2e363e", 1516044879, 0x1d0682e8 },
-    { 32254, "340d0f58bc3e28e65a441be2e8557243230632dc81ed8cf65f6f0677a18eb6cb", 1517147316, 0x1d00f989 },
-    { 41730, "c73361ab42465711db774796dce5686cefd8e44008638caaf02218a6ed4f75bb", 1520154717, 0x1d0512fa },
+    {     0, uint256("aec9203efaf5b35bdfc136477468830da578a1f5b21f4c59cf898dbb3fb17c33"), 1512973198, 0x1e10024c },
+    {     1, uint256("8852898d57539ff7e065b16149d2097a37b41617b963d6002a7a00ecd1eeadc8"), 1514518477, 0x1e10024c },
+    {  2608, uint256("22722429354af4f4ef8b9de55f0825611aa348d98fcb05d71e3c3adfaab4b94c"), 1514547220, 0x1e10024c },
+    {  5216, uint256("9726ea3e333ca64e6cffa5fa9fe0e27338634890b36ad09b75c9c04eee59d423"), 1514586552, 0x1e040093 },
+    { 10432, uint256("8c53c852c42c4ae1fad07c12cf8d259baf3ca873d24cac84cdf679d92dc6dcb6"), 1514770971, 0x1d100240 },
+    { 20864, uint256("12e76ce3468f0c2fa8d53458ac69967faf88cc1e1d43641cf3f2afb55d2e363e"), 1516044879, 0x1d0682e8 },
+    { 32254, uint256("340d0f58bc3e28e65a441be2e8557243230632dc81ed8cf65f6f0677a18eb6cb"), 1517147316, 0x1d00f989 },
+    { 41730, uint256("c73361ab42465711db774796dce5686cefd8e44008638caaf02218a6ed4f75bb"), 1520154717, 0x1d0512fa },
 };
 
 static const BRCheckPoint BRTestNetCheckpoints[] = {
-    { 0, "4966625a4b2851d9fdee139e56211a0d88575f59ed816ff5e6a63deb4e3e29a0", 1486949366, 0x1e0ffff0 },
+    { 0, uint256("4966625a4b2851d9fdee139e56211a0d88575f59ed816ff5e6a63deb4e3e29a0"), 1486949366, 0x1e0ffff0 },
 };
 
 static int BRMainNetVerifyDifficulty(const BRMerkleBlock *block, const BRSet *blockSet)
